Move integresql implementation details into src/internal/
Separate public API types (DatabaseTemplateId, DatabaseConfiguration,
IntegreSqlClient, NoSuchTemplate) from implementation code
(makeIntegreSqlClient, unsafeMakeDatabaseTemplateId, schema, errors)
by moving the latter into src/internal/integreSqlClient.ts. This
activates the existing "./internal/*": null export guard.
